fixed the tests that broke during the implementation of DeviceTime
