+++ lib/DeviceTime/include/DeviceTime.h
@@ -8,9 +8,11 @@
 #define __DEVICETIME_H__
 
 #include "DeviceTimeWithRTC.hpp"
-#include "DeviceTimeNoRTC.hpp"
+#include "DeviceTimeNoRTC.h"
 
 // default DeviceTimeClass does not integrate an RTC
-typedef DeviceTimeNoRTCClass DeviceTimeClass;
+// typedef DeviceTimeNoRTCClass DeviceTimeClass;
+
+
 
 #endif
