+++ test/EventManager/test_EventManager/test_EventManager.cpp
@@ -12,14 +12,14 @@ let convert2000To1970 = (timestamp) => {
 
 */
 #include "EventManager.h"
-#include "DeviceTime.h"
+// #include "DeviceTime.h"
 #include "mockModalLights.hpp"
 #include "DataStorageClass.h"
 
 #include "testEvents.h"
 #include "../../nativeMocksAndHelpers/mockConfig.h"
 #include "../../nativeMocksAndHelpers/mockStorageHAL.hpp"
-#include <iostream>
+// #include <iostream>
 
 void setUp(void) {
     // set stuff up here
